requirements.txt
hvac==0.7.2
argparse==1.4.0
pyyaml>=4.2b1
toml==0.10.0